member data store refactored to create its own graph client
