+++ src/routes/members/Member.svelte
@@ -1,14 +1,12 @@
 <script>
   export let router;
 
-  import { getClient } from '@urql/svelte';
-
   import MemberCore from './_MemberCore.svelte';
   import ProfileLayout from '../../elements/layouts/Profile.svelte';
-  import memberApi from '../../dataSources/api.that.tech/members';
+  import memberApi from '../../dataSources/api.that.tech/members/queries';
 
   const { member } = router.params;
-  const { queryMemberBySlug } = memberApi(getClient());
+  const { queryMemberBySlug } = memberApi();
 </script>
 
 <ProfileLayout>
