+++ src/routes/my/Profile.svelte
@@ -1,26 +1,25 @@
 <script>
-  import { getClient } from '@urql/svelte';
   import { navigateTo } from 'yrv';
-  import _ from 'lodash';
+  import { isNil, isEmpty } from 'lodash';
 
   import { ActionHeader } from '../../elements';
   import StackedLayout from '../../elements/layouts/StackedLayout.svelte';
   import Nav from '../../components/nav/interiorNav/Top.svelte';
   import ProfileForm from '../../components/my/ProfileForm.svelte';
 
-  import memberApi from '../../dataSources/api.that.tech/members.js';
+  import memberApi from '../../dataSources/api.that.tech/members/mutations';
   import { tagEvent } from '../../utilities/gtag';
   import logEvent from '../../utilities/eventTrack';
   import { user, thatProfile } from '../../utilities/security.js';
   import metaTagsStore from '../../store/metaTags';
 
-  const { createProfile, updateProfile } = memberApi(getClient());
+  const { createProfile, updateProfile } = memberApi();
 
   let isNewProfile;
   let currentProfile;
 
   $: if ($user || $thatProfile) {
-    if (!_.isNil($thatProfile) && !_.isEmpty($thatProfile)) {
+    if (!isNil($thatProfile) && !isEmpty($thatProfile)) {
       currentProfile = $thatProfile;
       isNewProfile = false;
     } else {
